+++ api/trace/span_context_test.go
@@ -176,7 +176,7 @@ func TestSpanContextIsSampled(t *testing.T) {
 			name: "sampled plus unused",
 			sc: trace.SpanContext{
 				TraceID:    trace.ID([16]byte{1}),
-				TraceFlags: trace.FlagsSampled | trace.FlagsUnused,
+				TraceFlags: trace.FlagsSampled | trace.FlagsUnset,
 			},
 			want: true,
 		}, {